+++ Scripts/Services/XmlSpawner/XmlUtils/XmlFind.cs
@@ -1169,14 +1169,14 @@ namespace Server.Mobiles
 
 
                 // add the page buttons
-                for (int i = 0; i < (int)(MaxEntries / MaxEntriesPerPage); i++)
+                for (int i = 0; i < MaxEntries / MaxEntriesPerPage; i++)
                 {
                     //AddButton( 38+i*30, 365, 2206, 2206, 0, GumpButtonType.Page, 1+i );
                     AddButton(418 + i * 25, height - 25, 0x8B1 + i, 0x8B1 + i, 0, GumpButtonType.Page, 1 + i);
                 }
 
                 // add the advance pageblock buttons
-                AddButton(415 + 25 * (int)(MaxEntries / MaxEntriesPerPage), height - 25, 0x15E1, 0x15E5, 201, GumpButtonType.Reply, 0); // block forward
+                AddButton(415 + 25 * (MaxEntries / MaxEntriesPerPage), height - 25, 0x15E1, 0x15E5, 201, GumpButtonType.Reply, 0); // block forward
                 AddButton(395, height - 25, 0x15E3, 0x15E7, 202, GumpButtonType.Reply, 0); // block backward
 
                 // add the displayfrom entry
@@ -1214,7 +1214,7 @@ namespace Server.Mobiles
 
                     SearchEntry e = (SearchEntry)m_SearchList[index];
 
-                    int page = (int)(i / MaxEntriesPerPage);
+                    int page = i / MaxEntriesPerPage;
 
                     if (i % MaxEntriesPerPage == 0)
                     {
